Column filtering, archive restoration, log fixes.
Users can now hide/show specific columns as they desire. Some columns are hidden by default on pages with an abundance of information that is likely to be unnecessary in most scenarios. However, users can opt in to see this information as they desire with the Hide/Show button at the top of any given grid.

The Archives page now allows Administrators to restore previously deleted orders. It also supports multiple row selection.

The LogEntry entity was updated to support logging of multiple-selection operations as well.