Priority UUIDs when creating users. Users can now be deleted as well. See commit details.
When creating new users, the system attempts to select one of twenty prioritized UUID options that are already registered with the CollaborationEngine. If all options are already taken, it leaves the default UUID generated by hibernate.

Administrators can now delete users from the system. If they are one of the original twenty, their UUID is preserved and will be given to one of the users created in the future.

No matter what, though, their UUID will be logged to Discord for record-keeping purposes if that feature is left enabled.