SQL update uses System.env()
On the ManageUsersView, the database variables are no-longer hardcoded into the source code.

Instead, they are fetched from the system's environment variables.

The reason we have a SQL update here is because we're recycling the same list of 20 basic UUIDs registered with the CollaborationEngine. To do this, we sometimes have to change a new User's dynamically generated UUID to one of our chosen, priority twenty. This cannot be done with JPA / Hibernate, but it can be done with a standard SQL update, which is what we're doing here.